+++ packages/service-core/src/routes/endpoints/sync-stream.ts
@@ -30,7 +30,10 @@ export const syncStreamed = routeDefinition({
     }
 
     const params: util.StreamingSyncRequest = payload.params;
-    const syncParams: SyncParameters = normalizeTokenParameters(payload.context.token_payload!.parameters ?? {});
+    const syncParams: SyncParameters = normalizeTokenParameters(
+      payload.context.token_payload!.parameters ?? {},
+      payload.params.parameters ?? {}
+    );
 
     const storage = system.storage;
     // Sanity check before we start the stream